schema is separated from form component, joined trigger messages into en dictionary
