+++ components/BusinessForm/index.ts
@@ -1,1 +1,2 @@
 export * from './BusinessForm';
+export * from './schema';
